add Middleware Admin
update AdminController
create Midleware Admin and add it on Kernel file
update admin route

+++ mint/routes/web.php
@@ -33,9 +33,9 @@ Route::get('/mentorac/{id}', 'MentorallconnectionController@index');
 Route::get('/disconnect/{id}', 'MentorallconnectionController@destroy');
 
 // ! Admin
-Route::get('/admin', 'AdminController@index');
-Route::put('/admin/update/{id}', 'AdminController@update');
-Route::delete('/admin/decline/{id}', 'AdminController@destroy');
+Route::get('/admin', 'AdminController@index')->middleware('admin');
+Route::put('/admin/update/{id}', 'AdminController@update')->middleware('admin');
+Route::delete('/admin/decline/{id}', 'AdminController@destroy')->middleware('admin');
 Route::get('/mentee/{id}', 'MenteeController@profile');
 
 //Route::get('/mentee/{id}', 'MenteeController@list');
